Make dispatcher thread safe.
Dispatcher is now I/O thread safe.
There is a mutex in `Register` method to ensure that insert a new method
doesn't create a race condition.
Another mutex is along each stored methods, they can then be called in
concurrency without problem.

The code come with a test to ensure that it works

Signed-off-by: Vasek - Tom C <tom.chauveau@epitech.eu>